+++ examples/AvroBlogExamples/Program.cs
@@ -39,7 +39,7 @@ namespace AvroBlogExample
             using (var schemaRegistry = new CachedSchemaRegistryClient(new SchemaRegistryConfig { SchemaRegistryUrl = schemaRegistryUrl }))
             using (var producer =
                 new ProducerBuilder<Null, GenericRecord>(new ProducerConfig { BootstrapServers = bootstrapServers })
-                    .SetValueSerializer(new AvroSerializer<GenericRecord>(schemaRegistry))
+                    .SetValueSerializer(new AsyncAvroSerializer<GenericRecord>(schemaRegistry))
                     .Build())
             {   
                 var logLevelSchema = (Avro.EnumSchema)Avro.Schema.Parse(
@@ -71,7 +71,7 @@ namespace AvroBlogExample
             using (var schemaRegistry = new CachedSchemaRegistryClient(new SchemaRegistryConfig { SchemaRegistryUrl = schemaRegistryUrl }))
             using (var producer =
                 new ProducerBuilder<Null, MessageTypes.LogMessage>(new ProducerConfig { BootstrapServers = bootstrapServers })
-                    .SetValueSerializer(new AvroSerializer<MessageTypes.LogMessage>(schemaRegistry))
+                    .SetValueSerializer(new AsyncAvroSerializer<MessageTypes.LogMessage>(schemaRegistry))
                     .Build())
             {
                 await producer.ProduceAsync("log-messages",
@@ -105,10 +105,10 @@ namespace AvroBlogExample
                 AutoOffsetReset = AutoOffsetReset.Earliest
             };
 
-            using (var schemaRegistry = new CachedSchemaRegistryClient( new SchemaRegistryConfig { SchemaRegistryUrl = schemaRegistryUrl }))
+            using (var schemaRegistry = new CachedSchemaRegistryClient(new SchemaRegistryConfig { SchemaRegistryUrl = schemaRegistryUrl }))
             using (var consumer =
                 new ConsumerBuilder<Null, MessageTypes.LogMessage>(consumerConfig)
-                    .SetValueDeserializer(new AvroDeserializer<MessageTypes.LogMessage>(schemaRegistry))
+                    .SetValueDeserializer(new AsyncAvroDeserializer<MessageTypes.LogMessage>(schemaRegistry).AsSyncOverAsync())
                     .Build())
             {
                 consumer.Subscribe("log-messages");